libdm: cast walk flags to uint64_t when logging.
Walk flags are ULL constants; cast the result to a uint64_t before
logging with a FMTx64 format specifier to avoid a compiler warning:

  warning: format ‘%lx’ expects argument of type ‘long unsigned int’,
  but argument 5 has type ‘long long unsigned int’

+++ libdm/libdm-stats.c
@@ -1399,7 +1399,7 @@ int dm_stats_walk_init(struct dm_stats *dms, uint64_t flags)
 
 	if (flags & ~DM_STATS_WALK_MASK) {
 		log_error("Unknown value in walk flags: 0x" FMTx64,
-			  flags & ~DM_STATS_WALK_MASK);
+			  (uint64_t) (flags & ~DM_STATS_WALK_MASK));
 		return 0;
 	}
 	dms->walk_flags = flags;
